feat:break annotator tools into separate submodule
