+++ src/ga4gh/vrs/extras/annotator/__init__.py
@@ -1,0 +1,1 @@
+"""Provide tools for annotating data with corresponding VRS objects and attributes."""
